requirements.txt
pymongo~=4.4.1
python-dotenv~=1.0.0
setuptools~=65.5.1
beautifulsoup4~=4.12.2
requests~=2.31.0
pytz~=2023.3
validators~=0.20.0